Task: Move vmcompose from boxEngine to vmcontroller
